https://github.com/DASISH/qddt-client/issues/770 https://github.com/DASISH/qddt-client/issues/769
Issue med Category children deserialzing/persisting

+++ src/main/kotlin/no/nsd/qddt/repository/PublicationStatusRepository.kt
@@ -16,7 +16,7 @@ import java.util.*
  */
 
 @RepositoryRestResource(path = "publicationstatus",  itemResourceRel = "PublicationStatus", excerptProjection = PublicationStatusListe::class)
-interface PublicationStatusRepository : JpaRepository<PublicationStatus, Long> {
+interface PublicationStatusRepository : JpaRepository<PublicationStatus, Int> {
 
     @RestResource(rel = "hierarchy", path = "/all")
     fun findAllByParentIdIsNull(): List<PublicationStatus>
